Moved functionality from cli.h
